Refactor main loop
This commit breaks up the code into two main functions:
- Update the letter grid and keyboard content with coloured letters etc.
  based on the player's guess.
- Reprint the player interface with the updated data.
The two functions' behaviour is governed by a dictionary in the top
scope, which holds the information about the current game state.
The main loop is moved to the top scope in the main program section and
rearranged in a way so that each of the functions described above as
well as the user input prompt is used exactly once in the source code.

Rationale: Make game logic easier to follow and also make the source
code more modular to make future enhancements easier to implement.