Fix bug in repeated letter evaluation
Likely fallout and/or oversight from
https://github.com/ShivamPandya/wordle/pull/9. Turning letters yellow
wasn't working correctly in cases when the first repeated letter occured
in the guess before its first occurrence in the correct answer and when
the second occurrence in the guess was placed correctly. E.g. when
guessing "LABEL" when the correct answer was "SKILL"